implement tail in xv6
Write a program that prints the last 10 lines of its input. If a filename is provided on the command line (i.e., tail FILE) then tail should open it, read and print the last 10 lines, and then close it. If no filename is provided, tail should read from standard input.

Support:
$ tail README
$ grep the README | tail
$ tail -2 README